(화,목) 창의적 종합 설계/쿠팡/requirements.txt
anyio @ file:///C:/ci/anyio_1644463705902/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///C:/ci/argon2-cffi-bindings_1644569848815/work
astor==0.8.1
attrdict==2.0.1
attrs @ file:///C:/b/abs_09s3y775ra/croot/attrs_1668696195628/work
Babel==2.14.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bce-python-sdk==0.9.5
beautifulsoup4==4.12.3
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
brotlipy==0.7.0
bs4==0.0.2
cachetools==5.3.3
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.9
debugpy @ file:///C:/ci/debugpy_1637091911212/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints @ file:///C:/ci/entrypoints_1649926659561/work
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastjsonschema @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_ebruxzvd08/croots/recipe/python-fastjsonschema_1661376484940/work
fire==0.6.0
Flask==2.2.5
flask-babel==3.1.0
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
future==1.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.31.2
imgaug==0.4.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1648562631189/work
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
ipykernel @ file:///C:/b/abs_21ykzkm7y_/croots/recipe/ipykernel_1662361803478/work
ipython @ file:///C:/Windows/TEMP/abs_45b5zb1l7q/croots/recipe/ipython_1659529855872/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///croot/ipywidgets_1701289330913/work
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644297241925/work
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///C:/b/abs_6ccs97j_l8/croot/jsonschema_1676558690963/work
jupyter @ file:///C:/Windows/TEMP/abs_56xfdi__li/croots/recipe/jupyter_1659349053177/work
jupyter-console @ file:///C:/b/abs_e8gx0csskd/croot/jupyter_console_1671541918848/work
jupyter-server @ file:///C:/b/abs_1cfi3__jl8/croot/jupyter_server_1671707636383/work
jupyter_client @ file:///C:/b/abs_d8fk_kz9zk/croot/jupyter_client_1676330195659/work
jupyter_core @ file:///C:/b/abs_b8r5lb5eus/croot/jupyter_core_1668084445779/work
jupyterlab @ file:///C:/b/abs_513jt6yy74/croot/jupyterlab_1675354138043/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
jupyterlab_server @ file:///C:/b/abs_d1z_g1swc8/croot/jupyterlab_server_1677153204814/work
kiwisolver==1.4.5
lmdb==1.4.1
lxml @ file:///C:/ci/lxml_1657509249535/work
MarkupSafe @ file:///C:/ci/markupsafe_1654508076077/work
matplotlib==3.5.3
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661934089728/work
mistune @ file:///C:/ci/mistune_1594373272338/work
nbclassic @ file:///C:/b/abs_d0_ze5q0j2/croot/nbclassic_1676902914817/work
nbclient @ file:///C:/ci/nbclient_1650308542913/work
nbconvert @ file:///C:/b/abs_4av3q4okro/croot/nbconvert_1668450658054/work
nbformat @ file:///C:/b/abs_85_3g7dkt4/croot/nbformat_1670352343720/work
nest-asyncio @ file:///C:/b/abs_3a_4jsjlqu/croot/nest-asyncio_1672387322800/work
networkx==2.6.3
notebook @ file:///C:/b/abs_ca13hqvuzw/croot/notebook_1668179888546/work
notebook_shim @ file:///C:/b/abs_ebfczttg6x/croot/notebook-shim_1668160590914/work
numpy==1.21.6
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging @ file:///C:/b/abs_cfsup8ur87/croot/packaging_1671697442297/work
paddleocr==2.7.0.2
paddlepaddle==2.5.2
pandas==1.3.5
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pdf2docx==0.5.8
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.5.0
pkgutil_resolve_name @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_81wm45v3kb/croots/recipe/pkgutil-resolve-name_1661463352381/work
ply==3.11
premailer==3.10.0
prometheus-client @ file:///C:/Windows/TEMP/abs_ab9nx8qb08/croots/recipe/prometheus_client_1659455104602/work
prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
protobuf==3.20.2
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pyclipper==1.3.0.post5
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.20.0
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyMuPDF==1.20.2
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing==3.1.2
PyQt5==5.15.7
PyQt5-sip @ file:///C:/Windows/Temp/abs_d7gmd2jg8i/croots/recipe/pyqt-split_1659273064801/work/pyqt_sip
pyrsistent @ file:///C:/ci/pyrsistent_1636093257833/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-docx==1.1.0
pytz @ file:///C:/b/abs_22fofvpn1x/croot/pytz_1671698059864/work
PyWavelets==1.3.0
pywin32==305.1
pywinpty @ file:///C:/b/abs_73vshmevwq/croot/pywinpty_1677609966356/work/target/wheels/pywinpty-2.0.10-cp37-none-win_amd64.whl
pyzmq @ file:///C:/ci/pyzmq_1657597757589/work
qtconsole @ file:///C:/b/abs_5bap7f8n0t/croot/qtconsole_1674008444833/work
QtPy @ file:///C:/ci/qtpy_1662015131946/work
rapidfuzz==3.4.0
rarfile==4.1
requests @ file:///C:/ci/requests_1657735288441/work
scikit-image==0.19.3
scipy==1.7.3
selenium==4.11.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
shapely==2.0.3
sip @ file:///C:/Windows/Temp/abs_b8fxd17m2u/croots/recipe/sip_1659012372737/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///C:/ci/sniffio_1614030522573/work
sortedcontainers==2.4.0
soupsieve==2.4.1
termcolor==2.3.0
terminado @ file:///C:/b/abs_25nakickad/croot/terminado_1671751845491/work
tifffile==2021.11.2
tinycss2 @ file:///C:/b/abs_52w5vfuaax/croot/tinycss2_1668168823131/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli @ file:///C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
tornado @ file:///C:/ci/tornado_1662476933490/work
tqdm==4.66.2
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
trio==0.22.2
trio-websocket==0.11.1
typing_extensions @ file:///C:/b/abs_89eui86zuq/croot/typing_extensions_1669923792806/work
urllib3 @ file:///C:/b/abs_9bcwxczrvm/croot/urllib3_1673575521331/work
visualdl==2.5.3
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///C:/ci/websocket-client_1614804293275/work
Werkzeug==2.2.3
widgetsnbextension @ file:///C:/ci/widgetsnbextension_1645009553925/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wsproto==1.2.0
zipp @ file:///C:/b/abs_b9jfdr908q/croot/zipp_1672387552360/work
